Fix parsing inconsistency mentioned in #23
